requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
distlib==0.3.7
Django==3.2.6
django-kendoui-backend==0.6
django-mssql-backend==2.8.1
djangorestframework==3.14.0
filelock==3.12.4
idna==3.4
json-utils==0.2
mssql-django==1.3
numpy==1.26.0
numpy-financial==1.0.0
pandas==2.1.0
platformdirs==3.11.0
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3
querystring-parser==1.2.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.24.6
